requirements.txt
langchain-openai
langchain-unstructured
langchain-community
watchdog
openai
streamlit
requests
python-dotenv
